wolframite/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-chess>=1.10.0
stockfish==0.1.0
